api/requirements.txt
flask==1.1.2
flask-restplus==0.13.0
pefile==2019.4.18
python-magic==0.4.18
requests==2.24.0
ssdeep==3.4
uwsgi==2.0.19.1
werkzeug==0.16.1
yara-python==4.0.2